Phase 4.2: seed the demo world
220 orders across 90 days, 90 verified-purchase reviews, 75 products, 14 shops,
41 users, offers, campaigns and notification history. db:reset completes clean in
about 3 seconds and prints row counts.

Deterministic by design: everything random comes from a seeded mulberry32 PRNG, so
every reset produces byte-identical data. Verified by fingerprinting the order set
across three consecutive resets. The presenter rehearses against specific orders,
so a reset that reshuffled the world mid-rehearsal would defeat the control panel.

Shape decisions that make the demo read as real:
- Order dates are recency-weighted (squared uniform draw) so the 30-day dashboard
  chart has a curve instead of a flat line
- Orders in the last two days carry a live tail of placed/accepted/ready, so the
  shop action queue is non-empty at demo start — verified across six shops
- Top five shops take a disproportionate share, giving the admin's top-shops
  report and revenue view something to show
- Ratings skew 4-5 with honest 2-3s and two reported reviews for the moderation
  queue; a wall of five stars reads as fabricated
- Featured-shops sits at 4 of 6 capacity, so the admin has visible unsold
  inventory rather than a fully booked slot

Two bugs found and fixed while verifying:

1. Every seeded notification stored a raw template key as its body. Event keys are
   dotted ('order.placed') and next-intl treats a dot as nesting, but the templates
   were flat keys containing dots, so nothing resolved. Only 'otp' worked because it
   has no dot. Worse, renderTemplate wrapped the lookup in try/catch — but
   createTranslator does not throw on a missing message, it returns the key path and
   reports via onError, so the Dari fallback never ran and the failure was silent.
   Templates are now nested, detection is via onError, and a miss logs a warning.
   This is precisely the failure the on-screen log panel exists to expose.

2. Aggregate fragments typed SQL<number> were returning STRINGS. Postgres
   count()/sum() are bigint and avg() is numeric; postgres.js maps both to strings
   to avoid precision loss. pendingShops().draftProductCount was "6", so `=== 6`
   failed and `count + 1` would have produced "61" in the admin UI. Every aggregate
   in fragments.ts is now cast ::int or ::float8, making the declared types honest.

Also: notification money is pre-formatted through lib/format.ts, so a Dari SMS
reads "؋ ۵۹٬۲۰۰" and an English one "AFN 18,500"; and the shop name in a
notification is localized per recipient rather than hardcoded Dari.

Verified: 41 checks via `npm run check:phase4`, reading through the real Phase 3
query modules so it doubles as an integration test of the data layer.

+++ lib/db/queries/fragments.ts
@@ -20,6 +20,12 @@ import { sql, type SQL } from 'drizzle-orm';
  *
  * Each fragment documents which table must be in the OUTER query's FROM clause,
  * because that is what the outer qualification binds to.
+ *
+ * EVERY aggregate IS EXPLICITLY CAST. Postgres count()/sum() return bigint and
+ * avg() returns numeric; postgres.js maps both to JavaScript STRINGS to avoid
+ * precision loss. Without ::int / ::float8 a fragment typed SQL<number> silently
+ * yields "6" instead of 6 — comparisons like `=== 6` fail and `count + 1` gives
+ * "61". Casting in SQL makes the declared TypeScript types honest at runtime.
  */
 
 /** Outer query must select FROM products. */
@@ -32,19 +38,19 @@ export const firstProductImagePath: SQL<string | null> = sql<string | null>`(
 
 /** Outer query must select FROM products. Visible reviews only. */
 export const productRatingAvg: SQL<number> = sql<number>`coalesce((
-  select avg(r.rating) from reviews r
+  select avg(r.rating)::float8 from reviews r
   where r.product_id = products.id and r.status = 'visible'
 ), 0)`;
 
 /** Outer query must select FROM products. */
 export const productReviewCount: SQL<number> = sql<number>`(
-  select count(*) from reviews r
+  select count(*)::int from reviews r
   where r.product_id = products.id and r.status = 'visible'
 )`;
 
 /** Outer query must select FROM products. */
 export const productWishlistCount: SQL<number> = sql<number>`(
-  select count(*) from wishlist_items wi
+  select count(*)::int from wishlist_items wi
   where wi.product_id = products.id
 )`;
 
@@ -54,33 +60,33 @@ export const productWishlistCount: SQL<number> = sql<number>`(
  * Outer query must select FROM shops.
  */
 export const shopRatingAvg: SQL<number> = sql<number>`coalesce((
-  select avg(r.rating) from reviews r
+  select avg(r.rating)::float8 from reviews r
   join products p on p.id = r.product_id
   where p.shop_id = shops.id and r.status = 'visible'
 ), 0)`;
 
 /** Outer query must select FROM shops. */
 export const shopReviewCount: SQL<number> = sql<number>`(
-  select count(*) from reviews r
+  select count(*)::int from reviews r
   join products p on p.id = r.product_id
   where p.shop_id = shops.id and r.status = 'visible'
 )`;
 
 /** Outer query must select FROM shops. Published products only. */
 export const shopPublishedProductCount: SQL<number> = sql<number>`(
-  select count(*) from products p
+  select count(*)::int from products p
   where p.shop_id = shops.id and p.status = 'published'
 )`;
 
 /** Outer query must select FROM shops. Counts drafts too, for the admin queue. */
 export const shopTotalProductCount: SQL<number> = sql<number>`(
-  select count(*) from products p
+  select count(*)::int from products p
   where p.shop_id = shops.id
 )`;
 
 /** Outer query must select FROM categories. Publicly visible products only. */
 export const categoryProductCount: SQL<number> = sql<number>`(
-  select count(*) from products p
+  select count(*)::int from products p
   join shops s on s.id = p.shop_id
   where p.category_id = categories.id
     and p.status = 'published' and s.status = 'approved'
@@ -88,13 +94,13 @@ export const categoryProductCount: SQL<number> = sql<number>`(
 
 /** Outer query must select FROM orders. */
 export const orderItemQuantity: SQL<number> = sql<number>`(
-  select coalesce(sum(oi.quantity), 0) from order_items oi
+  select coalesce(sum(oi.quantity), 0)::int from order_items oi
   where oi.order_id = orders.id
 )`;
 
 /** Outer query must select FROM orders. */
 export const orderShopCount: SQL<number> = sql<number>`(
-  select count(distinct oi.shop_id) from order_items oi
+  select count(distinct oi.shop_id)::int from order_items oi
   where oi.order_id = orders.id
 )`;
